+++ visualization/app/codeCharta/ui/dialog/dialog.module.ts
@@ -1,9 +1,0 @@
-import angular from "angular"
-import "angular-animate"
-import "angular-aria"
-import "angular-material"
-import "../../state/state.module"
-import { DialogService } from "./dialog.service"
-import camelCase from "lodash.camelcase"
-
-angular.module("app.codeCharta.ui.dialog", ["ngMaterial", "app.codeCharta.state"]).service(camelCase(DialogService.name), DialogService)